# packages/backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Scroblarr API
  version: 1.0.0
  description: |
    API documentation for Scroblarr - Media scrobbling service for Plex, Jellyfin, and Emby.

    This API allows you to manage users, sync media watching history across services, and configure integrations with TVTime and Trakt.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token (Plex or Jellyfin access token)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Global API key for API authentication
  schemas:
...
    Success:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
...
paths:
...
  /api/v1/webhooks/plex:
    post:
      summary: Plex webhook endpoint
      tags: [Webhooks]
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
          multipart/form-data:
            schema:
              type: object
              properties:
                payload:
                  type: string
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /api/v1/webhooks/jellyfin:
    post:
      summary: Jellyfin webhook endpoint
      tags: [Webhooks]
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Bad request
        '500':
          description: Internal server error
...